Move config observer to utils/configutil
Strips the stuttering Config prefix from the moved types: ConfigBuilder,
ConfigDefaulter, ConfigObserver and NewConfigObserver become Builder,
Defaulter, Observer and NewObserver.

The existing ConfigObserver interface in configutil/atomic.go would have
collided conceptually with the incoming Observer struct, so it is now
Observable, with ConfigObserver kept as an alias for external callers.

utils keeps aliases for all four moved names so dependent repos continue
to compile unchanged.
